third_party/WebKit/LayoutTests/push_messaging/push-messaging-api-surface-expected.txt

Tests the Push API surface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!navigator.push is true
PASS !!navigator.push.register is true
PASS navigator.push.register("senderId").constructor is Promise
PASS successfullyParsed is true

TEST COMPLETE
